Feature: Embedded with online help, so that user may always get the most up-to-date help. Bugfix:  Use long instead of int for volume. This is just a partial fix. We still need to do it for the rest.

+++ src/org/yccheok/jstock/gui/MutableStock.java
@@ -348,7 +348,8 @@ public class MutableStock {
         this.stock = s;        
     }
 
-    public void setVolume(int volume) {
+    // TODO: CRITICAL LONG BUG REVISED NEEDED.
+    public void setVolume(long volume) {
         final Stock s = new Stock(
                 stock.getCode(),
                 stock.getSymbol(),
